Configure the linter for project and remove linters.
