+++ Gemfile
@@ -1,0 +1,2 @@
+source 'https://rubygems.org'
+gem 'rubocop', '>= 1.0', '< 2.0'
