--- src/BaseDecorator.rb
@@ -1,11 +1,0 @@
-require_relative "Nameable.rb"
-
-class BaseDecorator < Nameable
-    def initialize(nameable)
-        @nameable = nameable
-    end
-
-    def correct_name
-        @nameable.correct_name
-    end
-end
