+++ src/Book.rb
@@ -1,10 +1,9 @@
 class Book
-    attr_accessor :title, :author, :rentals
-  
-    def initialize(title, author)
-      @title = title
-      @author = author
-      @rentals = []
-    end
+  attr_accessor :title, :author, :rentals
+
+  def initialize(title, author)
+    @title = title
+    @author = author
+    @rentals = []
   end
-  
+end
