--- src/CapitalizeDecorator.rb
@@ -1,7 +1,0 @@
-require_relative "BaseDecorator.rb"
-
-class CapitalizeDecorator < BaseDecorator
-    def correct_name
-        @nameable.correct_name.capitalize
-    end
-end
