+++ src/Classroom.rb
@@ -1,14 +1,13 @@
 class Classroom
-    attr_accessor :label, :students
-  
-    def initialize(label)
-      @label = label
-      @students = []
-    end
-  
-    def add_student(student)
-      @students << student
-      student.classroom = self
-    end
+  attr_accessor :label, :students
+
+  def initialize(label)
+    @label = label
+    @students = []
   end
-  
+
+  def add_student(student)
+    @students << student
+    student.classroom = self
+  end
+end
