+++ src/Nameable.rb
@@ -1,5 +1,5 @@
 class Nameable
-    def correct_name 
-        raise NotImplementedError, "You must implement the correct_name method in your class."
-    end
+  def correct_name
+    raise NotImplementedError, 'You must implement the correct_name method in your class.'
+  end
 end
