+++ src/Rental.rb
@@ -1,13 +1,12 @@
 class Rental
-    attr_accessor :date, :book, :person
-  
-    def initialize(date, book, person)
-      @date = date
-      @book = book
-      @person = person
-  
-      book.rentals << self
-      person.rentals << self
-    end
+  attr_accessor :date, :book, :person
+
+  def initialize(date, book, person)
+    @date = date
+    @book = book
+    @person = person
+
+    book.rentals << self
+    person.rentals << self
   end
-  
+end
